+++ tests/Integration.Tests/Integration.Tests.csproj
@@ -8,6 +8,7 @@
 
     <ItemGroup>
         <PackageReference Include="coverlet.collector" Version="6.0.2"/>
+        <PackageReference Include="JetBrains.Annotations" Version="2024.3.0"/>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1"/>
         <PackageReference Include="Moq" Version="4.20.72"/>
         <PackageReference Include="Moq.Contrib.HttpClient" Version="1.4.0"/>
