+++ tests/Integration.Tests/Service/Paper/PaperTradeCatchServiceTest.cs
@@ -1,0 +1,25 @@
+﻿using JetBrains.Annotations;
+using Moq;
+using Persistence;
+using Service.Interfaces;
+using Service.Paper;
+
+namespace Integration.Tests.Service.Paper;
+
+[TestSubject(typeof(PaperPortfolioService))]
+public class PaperTradeCatchServiceTest
+{
+    private readonly Mock<ApplicationDatabaseContext> _dbContext;
+    private readonly IPaperTradeCatchService _paperTradeCatchService;
+    private readonly Mock<IPaperPortfolioService> _portfolioService;
+
+    public PaperTradeCatchServiceTest(Mock<IPaperPortfolioService> portfolioService,
+        Mock<ApplicationDatabaseContext> dbContext)
+    {
+        _portfolioService = portfolioService;
+        _dbContext = dbContext;
+        _paperTradeCatchService = new PaperTradeCatchService(portfolioService.Object, dbContext.Object);
+    }
+
+    //TODO: Write integration tests when TestContainer is implemented
+}
