+++ tests/Service.Tests/Paper/PaperOrderServiceTest.cs
@@ -4,7 +4,6 @@ using Data.Models;
 using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 using Moq;
-using Persistence;
 using Service.Interfaces;
 using Service.Paper;
 
@@ -13,10 +12,10 @@ namespace Service.Tests.Paper;
 [TestSubject(typeof(PaperOrderService))]
 public class PaperOrderServiceTests
 {
-    private readonly PaperOrderService _paperOrderService;
     private readonly Mock<IBrokerDataService> _brokerDataService;
-    private readonly Mock<IPaperTradeCatchService> _paperTradeCatchService;
+    private readonly PaperOrderService _paperOrderService;
     private readonly Mock<IPaperPortfolioService> _paperPortfolioService;
+    private readonly Mock<IPaperTradeCatchService> _paperTradeCatchService;
 
     public PaperOrderServiceTests()
     {
@@ -24,7 +23,8 @@ public class PaperOrderServiceTests
         _paperPortfolioService = new Mock<IPaperPortfolioService>();
         _paperTradeCatchService = new Mock<IPaperTradeCatchService>();
         var logger = new Mock<ILogger<PaperOrderService>>();
-        _paperOrderService = new PaperOrderService(_paperPortfolioService.Object, _paperTradeCatchService.Object, _brokerDataService.Object, logger.Object);
+        _paperOrderService = new PaperOrderService(_paperPortfolioService.Object, _paperTradeCatchService.Object,
+            _brokerDataService.Object, logger.Object);
     }
 
     [Trait("Category", "Unit test")]
@@ -35,11 +35,13 @@ public class PaperOrderServiceTests
         var symbol = It.IsAny<string>();
         var quantity = It.IsAny<int>();
         var price = It.IsAny<decimal>();
-        
+
         var expectedResponse = new CurrentPriceResponse { Symbol = symbol, Price = price };
         _brokerDataService.Setup(api => api.GetCurrentPriceAsync(symbol)).ReturnsAsync(expectedResponse);
-        _paperTradeCatchService.Setup(api => api.CatchTrade(portfolioId, symbol, quantity, 45000.50m)).Returns(Task.CompletedTask);
-        _paperPortfolioService.Setup(api => api.PayReservedCash(portfolioId,quantity*45)).Returns(Task.CompletedTask);
+        _paperTradeCatchService.Setup(api => api.CatchTrade(portfolioId, symbol, quantity, 45000.50m))
+            .Returns(Task.CompletedTask);
+        _paperPortfolioService.Setup(api => api.PayReservedCash(portfolioId, quantity * 45))
+            .Returns(Task.CompletedTask);
 
         var result = await _paperOrderService.MarketOrder(portfolioId, symbol, quantity);
         Assert.NotNull(result);
@@ -47,7 +49,7 @@ public class PaperOrderServiceTests
         Assert.Equal(price, result.Price);
         Assert.Equal(quantity, result.Quantity);
     }
-    
+
     [Trait("Category", "Unit test")]
     [Fact(DisplayName = "Should unreserve cash and throw exception")]
     public async Task MarketOrder_ShouldUnreserveCashAndThrowException()
@@ -56,11 +58,11 @@ public class PaperOrderServiceTests
         var symbol = It.IsAny<string>();
         var quantity = It.IsAny<int>();
         var price = It.IsAny<decimal>();
-        
+
         var expectedResponse = new CurrentPriceResponse { Symbol = symbol, Price = price };
         _brokerDataService.Setup(api => api.GetCurrentPriceAsync(symbol)).ReturnsAsync(expectedResponse);
         _paperTradeCatchService.Setup(api => api.CatchTrade(portfolioId, symbol, quantity, price)).Throws<Exception>();
-        
+
         await Assert.ThrowsAsync<Exception>(() => _paperOrderService.MarketOrder(portfolioId, symbol, quantity));
         _paperPortfolioService.Verify(api => api.UnreserveCash(portfolioId, It.IsAny<decimal>()), Times.Once);
     }
